Shuffle structs into correct package
